fix: moved tests inside pallet-ethereum src and replaced crate with pallet_3vm_evm from circuit_mock_runtime
