+++ pallets/3vm/ethereum/src/lib.rs
@@ -552,7 +552,7 @@ impl<T: Config> Pallet<T> {
         builder.build()
     }
 
-    fn apply_validated_transaction(
+    pub fn apply_validated_transaction(
         source: H160,
         transaction: Transaction,
     ) -> DispatchResultWithPostInfo {
