+++ pallets/3vm/ethereum/src/tests/eip2930.rs
@@ -16,16 +16,27 @@
 //! Consensus extension module tests for BABE consensus.
 
 use super::*;
+use circuit_mock_runtime::{
+    contract_address, new_test_ext, new_test_ext_with_initial_balance, storage_address,
+    AccountMapping, Balances, EIP2930UnsignedTransaction, Ethereum, Evm, ExtBuilder, Runtime,
+    RuntimeCall, RuntimeEvent, RuntimeOrigin, SignedExtra, System, ALICE, BOB,
+};
 use evm::{ExitReason, ExitRevert, ExitSucceed};
 use fp_ethereum::ValidatedTransaction;
-use frame_support::{dispatch::DispatchClass, traits::Get, weights::Weight};
-use pallet_3vm_evm::{AddressMapping, GasWeightMapping};
+use fp_self_contained::CheckedExtrinsic;
+use frame_support::{
+    assert_err,
+    dispatch::{DispatchClass, GetDispatchInfo},
+    weights::Weight,
+};
+use pallet_3vm_evm::AddressMapping;
+use sp_core::{H256, U256};
+use std::str::FromStr;
 
-fn eip1559_erc20_creation_unsigned_transaction() -> EIP1559UnsignedTransaction {
-    EIP1559UnsignedTransaction {
+fn eip2930_erc20_creation_unsigned_transaction() -> EIP2930UnsignedTransaction {
+    EIP2930UnsignedTransaction {
         nonce: U256::zero(),
-        max_priority_fee_per_gas: U256::from(1),
-        max_fee_per_gas: U256::from(1),
+        gas_price: U256::from(1),
         gas_limit: U256::from(0x100000),
         action: ethereum::TransactionAction::Create,
         value: U256::zero(),
@@ -33,86 +44,8 @@ fn eip1559_erc20_creation_unsigned_transaction() -> EIP1559UnsignedTransaction {
     }
 }
 
-fn eip1559_erc20_creation_transaction(account: &AccountInfo) -> Transaction {
-    eip1559_erc20_creation_unsigned_transaction().sign(&account.private_key, None)
-}
-
-#[test]
-fn transaction_with_max_extrinsic_gas_limit_should_success_pre_dispatch() {
-    let (pairs, mut ext) = new_test_ext_with_initial_balance(2, 10_000_000_000_000);
-    let alice = &pairs[0];
-    let bob = &pairs[1];
-
-    let limits: frame_system::limits::BlockWeights =
-        <Test as frame_system::Config>::BlockWeights::get();
-    let max_extrinsic = limits.get(DispatchClass::Normal).max_extrinsic.unwrap();
-    let max_extrinsic_gas =
-        <Test as pallet_3vm_evm::Config>::GasWeightMapping::weight_to_gas(max_extrinsic);
-
-    ext.execute_with(|| {
-        let transaction = EIP1559UnsignedTransaction {
-            nonce: U256::zero(),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
-            gas_limit: U256::from(max_extrinsic_gas),
-            action: ethereum::TransactionAction::Call(bob.address),
-            value: U256::from(1),
-            input: Default::default(),
-        }
-        .sign(&alice.private_key, None);
-
-        let call = crate::Call::<Test>::transact { transaction };
-        let source = call.check_self_contained().unwrap().unwrap();
-        let extrinsic = CheckedExtrinsic::<u64, _, SignedExtra, _> {
-            signed: fp_self_contained::CheckedSignature::SelfContained(source),
-            function: RuntimeCall::Ethereum(call.clone()),
-        };
-
-        assert_ok!(call
-            .pre_dispatch_self_contained(&source, &extrinsic.get_dispatch_info(), 0)
-            .unwrap());
-    });
-}
-
-#[test]
-fn transaction_with_gas_limit_greater_than_max_extrinsic_should_fail_pre_dispatch() {
-    let (pairs, mut ext) = new_test_ext_with_initial_balance(2, 10_000_000_000_000);
-    let alice = &pairs[0];
-    let bob = &pairs[1];
-
-    let limits: frame_system::limits::BlockWeights =
-        <Test as frame_system::Config>::BlockWeights::get();
-    let max_extrinsic = limits.get(DispatchClass::Normal).max_extrinsic.unwrap();
-    let base_extrinsic = limits.get(DispatchClass::Normal).base_extrinsic;
-    let max_extrinsic_gas = <Test as pallet_3vm_evm::Config>::GasWeightMapping::weight_to_gas(
-        max_extrinsic + base_extrinsic,
-    );
-
-    ext.execute_with(|| {
-        let transaction = EIP1559UnsignedTransaction {
-            nonce: U256::zero(),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
-            gas_limit: U256::from(max_extrinsic_gas + 1),
-            action: ethereum::TransactionAction::Call(bob.address),
-            value: U256::from(1),
-            input: Default::default(),
-        }
-        .sign(&alice.private_key, None);
-
-        let call = crate::Call::<Test>::transact { transaction };
-        let source = call.check_self_contained().unwrap().unwrap();
-        let extrinsic = CheckedExtrinsic::<u64, _, SignedExtra, _> {
-            signed: fp_self_contained::CheckedSignature::SelfContained(source),
-            function: RuntimeCall::Ethereum(call.clone()),
-        };
-
-        assert_err!(
-            call.pre_dispatch_self_contained(&source, &extrinsic.get_dispatch_info(), 0)
-                .unwrap(),
-            InvalidTransaction::ExhaustsResources
-        );
-    });
+fn eip2930_erc20_creation_transaction(account: &AccountInfo) -> Transaction {
+    eip2930_erc20_creation_unsigned_transaction().sign(&account.private_key, None)
 }
 
 #[test]
@@ -121,9 +54,14 @@ fn transaction_should_increment_nonce() {
     let alice = &pairs[0];
 
     ext.execute_with(|| {
-        let t = eip1559_erc20_creation_transaction(alice);
+        let t = eip2930_erc20_creation_transaction(alice);
         assert_ok!(Ethereum::execute(alice.address, &t, None,));
-        assert_eq!(EVM::account_basic(&alice.address).0.nonce, U256::from(1));
+        assert_eq!(
+            pallet_3vm_evm::Pallet::<Runtime>::account_basic(&alice.address)
+                .0
+                .nonce,
+            U256::from(1)
+        );
     });
 }
 
@@ -133,19 +71,21 @@ fn transaction_without_enough_gas_should_not_work() {
     let alice = &pairs[0];
 
     ext.execute_with(|| {
-        let mut transaction = eip1559_erc20_creation_transaction(alice);
+        let mut transaction = eip2930_erc20_creation_transaction(alice);
         match &mut transaction {
-            Transaction::EIP1559(t) => t.max_fee_per_gas = U256::from(11_000_000),
+            Transaction::EIP2930(t) => t.gas_price = U256::from(11_000_000),
             _ => {},
         }
 
-        let call = crate::Call::<Test>::transact { transaction };
+        let call =
+            circuit_mock_runtime::pallet_3vm_ethereum::Call::<Runtime>::transact { transaction };
         let source = call.check_self_contained().unwrap().unwrap();
         let extrinsic = CheckedExtrinsic::<u64, _, SignedExtra, _> {
             signed: fp_self_contained::CheckedSignature::SelfContained(source),
             function: RuntimeCall::Ethereum(call.clone()),
         };
         let dispatch_info = extrinsic.get_dispatch_info();
+
         assert_err!(
             call.validate_self_contained(&source, &dispatch_info, 0)
                 .unwrap(),
@@ -161,10 +101,11 @@ fn transaction_with_to_low_nonce_should_not_work() {
 
     ext.execute_with(|| {
         // nonce is 0
-        let mut transaction = eip1559_erc20_creation_unsigned_transaction();
+        let mut transaction = eip2930_erc20_creation_unsigned_transaction();
         transaction.nonce = U256::from(1);
+
         let signed = transaction.sign(&alice.private_key, None);
-        let call = crate::Call::<Test>::transact {
+        let call = circuit_mock_runtime::pallet_3vm_ethereum::Call::<Runtime>::transact {
             transaction: signed,
         };
         let source = call.check_self_contained().unwrap().unwrap();
@@ -184,14 +125,15 @@ fn transaction_with_to_low_nonce_should_not_work() {
                 .build()
         );
 
-        let t = eip1559_erc20_creation_transaction(alice);
+        let t = eip2930_erc20_creation_transaction(alice);
 
         // nonce is 1
         assert_ok!(Ethereum::execute(alice.address, &t, None,));
+
         transaction.nonce = U256::from(0);
 
         let signed2 = transaction.sign(&alice.private_key, None);
-        let call2 = crate::Call::<Test>::transact {
+        let call2 = circuit_mock_runtime::pallet_3vm_ethereum::Call::<Runtime>::transact {
             transaction: signed2,
         };
         let source2 = call2.check_self_contained().unwrap().unwrap();
@@ -215,11 +157,11 @@ fn transaction_with_to_hight_nonce_should_fail_in_block() {
     let alice = &pairs[0];
 
     ext.execute_with(|| {
-        let mut transaction = eip1559_erc20_creation_unsigned_transaction();
+        let mut transaction = eip2930_erc20_creation_unsigned_transaction();
         transaction.nonce = U256::one();
 
         let signed = transaction.sign(&alice.private_key, None);
-        let call = crate::Call::<Test>::transact {
+        let call = circuit_mock_runtime::pallet_3vm_ethereum::Call::<Runtime>::transact {
             transaction: signed,
         };
         let source = call.check_self_contained().unwrap().unwrap();
@@ -229,7 +171,7 @@ fn transaction_with_to_hight_nonce_should_fail_in_block() {
         };
         let dispatch_info = extrinsic.get_dispatch_info();
         assert_err!(
-            extrinsic.apply::<Test>(&dispatch_info, 0),
+            extrinsic.apply::<Runtime>(&dispatch_info, 0),
             TransactionValidityError::Invalid(InvalidTransaction::Future)
         );
     });
@@ -242,9 +184,10 @@ fn transaction_with_invalid_chain_id_should_fail_in_block() {
 
     ext.execute_with(|| {
         let transaction =
-            eip1559_erc20_creation_unsigned_transaction().sign(&alice.private_key, Some(1));
+            eip2930_erc20_creation_unsigned_transaction().sign(&alice.private_key, Some(1));
 
-        let call = crate::Call::<Test>::transact { transaction };
+        let call =
+            circuit_mock_runtime::pallet_3vm_ethereum::Call::<Runtime>::transact { transaction };
         let source = call.check_self_contained().unwrap().unwrap();
         let extrinsic = CheckedExtrinsic::<_, _, SignedExtra, _> {
             signed: fp_self_contained::CheckedSignature::SelfContained(source),
@@ -252,7 +195,7 @@ fn transaction_with_invalid_chain_id_should_fail_in_block() {
         };
         let dispatch_info = extrinsic.get_dispatch_info();
         assert_err!(
-            extrinsic.apply::<Test>(&dispatch_info, 0),
+            extrinsic.apply::<Runtime>(&dispatch_info, 0),
             TransactionValidityError::Invalid(InvalidTransaction::Custom(
                 fp_ethereum::TransactionValidationError::InvalidChainId as u8,
             ))
@@ -268,11 +211,11 @@ fn contract_constructor_should_get_executed() {
     let alice_storage_address = storage_address(alice.address, H256::zero());
 
     ext.execute_with(|| {
-        let t = eip1559_erc20_creation_transaction(alice);
+        let t = eip2930_erc20_creation_transaction(alice);
 
         assert_ok!(Ethereum::execute(alice.address, &t, None,));
         assert_eq!(
-            pallet_3vm_evm::AccountStorages::<Test>::get(erc20_address, alice_storage_address),
+            pallet_3vm_evm::AccountStorages::<Runtime>::get(erc20_address, alice_storage_address),
             H256::from_str("ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff")
                 .unwrap()
         )
@@ -290,13 +233,13 @@ fn source_should_be_derived_from_signature() {
     ext.execute_with(|| {
         Ethereum::transact(
             RawOrigin::EthereumTransaction(alice.address).into(),
-            eip1559_erc20_creation_transaction(alice),
+            eip2930_erc20_creation_transaction(alice),
         )
         .expect("Failed to execute transaction");
 
         // We verify the transaction happened with alice account.
         assert_eq!(
-            pallet_3vm_evm::AccountStorages::<Test>::get(erc20_address, alice_storage_address),
+            pallet_3vm_evm::AccountStorages::<Runtime>::get(erc20_address, alice_storage_address),
             H256::from_str("ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff")
                 .unwrap()
         )
@@ -311,10 +254,10 @@ fn contract_should_be_created_at_given_address() {
     let erc20_address = contract_address(alice.address, 0);
 
     ext.execute_with(|| {
-        let t = eip1559_erc20_creation_transaction(alice);
+        let t = eip2930_erc20_creation_transaction(alice);
         assert_ok!(Ethereum::execute(alice.address, &t, None,));
         assert_ne!(
-            pallet_3vm_evm::AccountCodes::<Test>::get(erc20_address).len(),
+            pallet_3vm_evm::AccountCodes::<Runtime>::get(erc20_address).len(),
             0
         );
     });
@@ -328,7 +271,7 @@ fn transaction_should_generate_correct_gas_used() {
     let expected_gas = U256::from(894198);
 
     ext.execute_with(|| {
-        let t = eip1559_erc20_creation_transaction(alice);
+        let t = eip2930_erc20_creation_transaction(alice);
         let (_, _, info) = Ethereum::execute(alice.address, &t, None).unwrap();
 
         match info {
@@ -346,10 +289,9 @@ fn call_should_handle_errors() {
     let alice = &pairs[0];
 
     ext.execute_with(|| {
-        let t = EIP1559UnsignedTransaction {
+        let t = EIP2930UnsignedTransaction {
             nonce: U256::zero(),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
+            gas_price: U256::from(1),
             gas_limit: U256::from(0x100000),
             action: ethereum::TransactionAction::Create,
             value: U256::zero(),
@@ -362,10 +304,9 @@ fn call_should_handle_errors() {
         let foo = hex::decode("c2985578").unwrap();
         let bar = hex::decode("febb0f7e").unwrap();
 
-        let t2 = EIP1559UnsignedTransaction {
+        let t2 = EIP2930UnsignedTransaction {
             nonce: U256::from(1),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
+            gas_price: U256::from(1),
             gas_limit: U256::from(0x100000),
             action: TransactionAction::Call(H160::from_slice(&contract_address)),
             value: U256::zero(),
@@ -386,10 +327,9 @@ fn call_should_handle_errors() {
             CallOrCreateInfo::Create(_) => panic!("expected call info"),
         }
 
-        let t3 = EIP1559UnsignedTransaction {
+        let t3 = EIP2930UnsignedTransaction {
             nonce: U256::from(2),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
+            gas_price: U256::from(1),
             gas_limit: U256::from(0x100000),
             action: TransactionAction::Call(H160::from_slice(&contract_address)),
             value: U256::zero(),
@@ -410,10 +350,9 @@ fn event_extra_data_should_be_handle_properly() {
     ext.execute_with(|| {
         System::set_block_number(1);
 
-        let t = EIP1559UnsignedTransaction {
+        let t = EIP2930UnsignedTransaction {
             nonce: U256::zero(),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
+            gas_price: U256::from(1),
             gas_limit: U256::from(0x100000),
             action: ethereum::TransactionAction::Create,
             value: U256::zero(),
@@ -426,10 +365,9 @@ fn event_extra_data_should_be_handle_properly() {
         let foo = hex::decode("c2985578").unwrap();
         let bar = hex::decode("febb0f7e").unwrap();
 
-        let t2 = EIP1559UnsignedTransaction {
+        let t2 = EIP2930UnsignedTransaction {
             nonce: U256::from(1),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
+            gas_price: U256::from(1),
             gas_limit: U256::from(0x100000),
             action: TransactionAction::Call(H160::from_slice(&contract_address)),
             value: U256::zero(),
@@ -439,21 +377,22 @@ fn event_extra_data_should_be_handle_properly() {
 
         // calling foo
         assert_ok!(Ethereum::apply_validated_transaction(alice.address, t2,));
-        System::assert_last_event(RuntimeEvent::Ethereum(Event::Executed {
-            from: alice.address,
-            to: H160::from_slice(&contract_address),
-            transaction_hash: H256::from_str(
-                "0x66036e60bafa1fabb847679d86d1f0a027ae52afde7690aaadf6f81c7e8ecda2",
-            )
-            .unwrap(),
-            exit_reason: ExitReason::Succeed(ExitSucceed::Returned),
-            extra_data: vec![],
-        }));
+        System::assert_last_event(RuntimeEvent::Ethereum(
+            circuit_mock_runtime::pallet_3vm_ethereum::Event::Executed {
+                from: alice.address,
+                to: H160::from_slice(&contract_address),
+                transaction_hash: H256::from_str(
+                    "0xd953990ba4a99074c5b8c7a9f97546ffbf781815acc2dcc2719ac3d0c8955a00",
+                )
+                .unwrap(),
+                exit_reason: ExitReason::Succeed(ExitSucceed::Returned),
+                extra_data: vec![],
+            },
+        ));
 
-        let t3 = EIP1559UnsignedTransaction {
+        let t3 = EIP2930UnsignedTransaction {
             nonce: U256::from(2),
-            max_priority_fee_per_gas: U256::from(1),
-            max_fee_per_gas: U256::from(1),
+            gas_price: U256::from(1),
             gas_limit: U256::from(0x100000),
             action: TransactionAction::Call(H160::from_slice(&contract_address)),
             value: U256::zero(),
@@ -463,16 +402,18 @@ fn event_extra_data_should_be_handle_properly() {
 
         // calling bar revert
         assert_ok!(Ethereum::apply_validated_transaction(alice.address, t3,));
-        System::assert_last_event(RuntimeEvent::Ethereum(Event::Executed {
-            from: alice.address,
-            to: H160::from_slice(&contract_address),
-            transaction_hash: H256::from_str(
-                "0xf14cd5e6d0535a8407339dabce26616a3f19c2048937450859ea65d5201fb7e4",
-            )
-            .unwrap(),
-            exit_reason: ExitReason::Revert(ExitRevert::Reverted),
-            extra_data: b"very_long_error_msg_that_we_ex".to_vec(),
-        }));
+        System::assert_last_event(RuntimeEvent::Ethereum(
+            circuit_mock_runtime::pallet_3vm_ethereum::Event::Executed {
+                from: alice.address,
+                to: H160::from_slice(&contract_address),
+                transaction_hash: H256::from_str(
+                    "0xbcb3fd14507b3aba8a124b007d959bf18d59f87c6d6c036e8feac224ec5fabeb",
+                )
+                .unwrap(),
+                exit_reason: ExitReason::Revert(ExitRevert::Reverted),
+                extra_data: b"very_long_error_msg_that_we_ex".to_vec(),
+            },
+        ));
     });
 }
 
@@ -489,20 +430,20 @@ fn self_contained_transaction_with_extra_gas_should_adjust_weight_with_post_disp
     .base_extrinsic;
 
     ext.execute_with(|| {
-        let mut transaction = eip1559_erc20_creation_unsigned_transaction();
+        let mut transaction = eip2930_erc20_creation_unsigned_transaction();
         transaction.gas_limit = 9_000_000.into();
         let signed = transaction.sign(&alice.private_key, None);
-        let call = crate::Call::<Test>::transact {
+        let call = circuit_mock_runtime::pallet_3vm_ethereum::Call::<Runtime>::transact {
             transaction: signed,
         };
         let source = call.check_self_contained().unwrap().unwrap();
-        let extrinsic = CheckedExtrinsic::<_, _, frame_system::CheckWeight<Test>, _> {
+        let extrinsic = CheckedExtrinsic::<_, _, frame_system::CheckWeight<Runtime>, _> {
             signed: fp_self_contained::CheckedSignature::SelfContained(source),
             function: RuntimeCall::Ethereum(call),
         };
         let dispatch_info = extrinsic.get_dispatch_info();
         let post_dispatch_weight = extrinsic
-            .apply::<Test>(&dispatch_info, 0)
+            .apply::<Runtime>(&dispatch_info, 0)
             .unwrap()
             .unwrap()
             .actual_weight
@@ -510,7 +451,7 @@ fn self_contained_transaction_with_extra_gas_should_adjust_weight_with_post_disp
 
         let expected_weight = base_extrinsic_weight.saturating_add(post_dispatch_weight);
         let actual_weight =
-            *frame_system::Pallet::<Test>::block_weight().get(DispatchClass::Normal);
+            *frame_system::Pallet::<Runtime>::block_weight().get(DispatchClass::Normal);
         assert_eq!(
             expected_weight,
             actual_weight,
@@ -526,15 +467,14 @@ fn validated_transaction_apply_zero_gas_price_works() {
     let alice = &pairs[0];
     let bob = &pairs[1];
     let substrate_alice =
-        <Test as pallet_3vm_evm::Config>::AddressMapping::into_account_id(alice.address);
+        <Runtime as pallet_3vm_evm::Config>::AddressMapping::into_account_id(alice.address);
     let substrate_bob =
-        <Test as pallet_3vm_evm::Config>::AddressMapping::into_account_id(bob.address);
+        <Runtime as pallet_3vm_evm::Config>::AddressMapping::into_account_id(bob.address);
 
     ext.execute_with(|| {
-        let transaction = EIP1559UnsignedTransaction {
+        let transaction = EIP2930UnsignedTransaction {
             nonce: U256::zero(),
-            max_priority_fee_per_gas: U256::zero(),
-            max_fee_per_gas: U256::zero(),
+            gas_price: U256::zero(),
             gas_limit: U256::from(21_000),
             action: ethereum::TransactionAction::Call(bob.address),
             value: U256::from(100),
@@ -542,10 +482,12 @@ fn validated_transaction_apply_zero_gas_price_works() {
         }
         .sign(&alice.private_key, None);
 
-        assert_ok!(crate::ValidatedTransaction::<Test>::apply(
-            alice.address,
-            transaction
-        ));
+        assert_ok!(
+            circuit_mock_runtime::pallet_3vm_ethereum::ValidatedTransaction::<Runtime>::apply(
+                alice.address,
+                transaction
+            )
+        );
         // Alice didn't pay fees, transfer 100 to Bob.
         assert_eq!(Balances::free_balance(&substrate_alice), 900);
         // Bob received 100 from Alice.
@@ -555,14 +497,15 @@ fn validated_transaction_apply_zero_gas_price_works() {
 
 #[test]
 fn proof_size_weight_limit_validation_works() {
+    use pallet_3vm_evm::GasWeightMapping;
+
     let (pairs, mut ext) = new_test_ext(1);
     let alice = &pairs[0];
 
     ext.execute_with(|| {
-        let mut tx = EIP1559UnsignedTransaction {
+        let mut tx = EIP2930UnsignedTransaction {
             nonce: U256::from(2),
-            max_priority_fee_per_gas: U256::zero(),
-            max_fee_per_gas: U256::from(1),
+            gas_price: U256::from(1),
             gas_limit: U256::from(0x100000),
             action: ethereum::TransactionAction::Call(alice.address),
             value: U256::from(1),
@@ -573,7 +516,7 @@ fn proof_size_weight_limit_validation_works() {
         tx.gas_limit = U256::from(gas_limit);
 
         let weight_limit =
-            <Test as pallet_3vm_evm::Config>::GasWeightMapping::gas_to_weight(gas_limit, true);
+            <Runtime as pallet_3vm_evm::Config>::GasWeightMapping::gas_to_weight(gas_limit, true);
 
         // Gas limit cannot afford the extra byte and thus is expected to exhaust.
         tx.input = vec![0u8; (weight_limit.proof_size() + 1) as usize];
